Exercise 9 Completed
There's a bug in the code so that if you run the section that uses left.join to join "Items" to the "LexicalData" data frame, it'll keep left-joining as many times as you please. I think that also screws up your data in that it'll start restricting the size of the "LexicalData" df to a 6xn(# times code was ran + appended variables) size.